requirements.txt
flask
flask-wtf
flask-bootstrap
email-validator
flask-sqlalchemy
mariadb
flask-login
imdbpy
pytest
flask-bcrypt
flask-mail
pyjwt
